--- src/main/java/com/codefusion/wasbackend/Account/Model/AccountEntity.java
@@ -1,26 +1,0 @@
-package com.codefusion.wasbackend.Account.Model;
-
-
-import com.codefusion.wasbackend.User.Model.UserEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "account")
-public class AccountEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @OneToOne(cascade = CascadeType.ALL)
-    @JoinColumn(name = "user_id", referencedColumnName = "id")
-    private UserEntity user;
-
-    private String username;
-    private String password;
-}
