--- src/main/java/com/codefusion/wasbackend/Store/Model/StoreEntity.java
@@ -1,31 +1,0 @@
-package com.codefusion.wasbackend.Store.Model;
-
-import com.codefusion.wasbackend.Product.Model.ProductEntity;
-import com.codefusion.wasbackend.User.Model.UserEntity;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-@Entity
-@Table(name = "store")
-public class StoreEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @NotBlank(message = "Name cannot be empty")
-    @Column(name = "name")
-    private String name;
-
-    @ManyToMany(fetch = FetchType.EAGER)
-    @JoinTable(name = "user_stores",
-            joinColumns = @JoinColumn(name = "store_id"),
-            inverseJoinColumns = @JoinColumn(name = "user_id"))
-    private List<UserEntity> user;
-
-    @OneToMany(mappedBy = "store", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<ProductEntity> products;
-}
