--- src/main/java/com/codefusion/wasbackend/User/Role.java
@@ -1,5 +1,0 @@
-package com.codefusion.wasbackend.User;
-
-public enum Role {
-    ADMIN,BOSS,EMPLOYEE
-}
