--- src/main/java/com/codefusion/wasbackend/User/Service/UserService.java
@@ -1,14 +1,0 @@
-package com.codefusion.wasbackend.User.Service;
-
-import com.codefusion.wasbackend.User.Repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@RequiredArgsConstructor
-@Service
-public class UserService {
-
-    private final UserRepository userRepository;
-
-
-}
