+++ src/main/java/com/codefusion/wasbackend/account/model/AccountEntity.java
@@ -1,0 +1,26 @@
+package com.codefusion.wasbackend.account.model;
+
+
+import com.codefusion.wasbackend.user.model.UserEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(name = "account")
+public class AccountEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @OneToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "user_id", referencedColumnName = "id")
+    private UserEntity user;
+
+    private String username;
+    private String password;
+}
