+++ src/main/java/com/codefusion/wasbackend/account/repository/AccountRepository.java
@@ -1,0 +1,9 @@
+package com.codefusion.wasbackend.account.repository;
+
+import com.codefusion.wasbackend.account.model.AccountEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AccountRepository extends JpaRepository<AccountEntity, Long> {
+}
