+++ src/main/java/com/codefusion/wasbackend/store/model/StoreEntity.java
@@ -1,0 +1,31 @@
+package com.codefusion.wasbackend.store.model;
+
+import com.codefusion.wasbackend.product.model.ProductEntity;
+import com.codefusion.wasbackend.user.model.UserEntity;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+@Entity
+@Table(name = "store")
+public class StoreEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @NotBlank(message = "Name cannot be empty")
+    @Column(name = "name")
+    private String name;
+
+    @ManyToMany(fetch = FetchType.EAGER)
+    @JoinTable(name = "user_stores",
+            joinColumns = @JoinColumn(name = "store_id"),
+            inverseJoinColumns = @JoinColumn(name = "user_id"))
+    private List<UserEntity> user;
+
+    @OneToMany(mappedBy = "store", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<ProductEntity> products;
+}
