+++ src/main/java/com/codefusion/wasbackend/user/Role.java
@@ -1,0 +1,5 @@
+package com.codefusion.wasbackend.user;
+
+public enum Role {
+    ADMIN,BOSS,EMPLOYEE
+}
