+++ src/main/java/com/codefusion/wasbackend/user/service/UserService.java
@@ -1,0 +1,14 @@
+package com.codefusion.wasbackend.user.service;
+
+import com.codefusion.wasbackend.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@RequiredArgsConstructor
+@Service
+public class UserService {
+
+    private final UserRepository userRepository;
+
+
+}
